Initial commit of some changes to the resolver for XML catalogs
Adding a configurable factory that can control what resolver is
used as the "user-chosen" resolver for XML catalog usage.

The issue surrounds the fact that the default CatalogResolver
doesn't seem to handle certain system catalog entries correctly.
This enables you to substitute in the Xerces resolver.jar classes
instead, which is still not without issue but at least has
workarounds instead of just failing.

More work will be done on testing and bug fixing.